increase keypad reliability by cheating
because of an instable voltage when pressing key 7
the key was sometimes detected as a 10 because of
a too high voltage. this is solved by "cheating":
a can no longer be detected because all high voltages
are mapped to 7

this voltage problem occured in 53/62 = 85% of
attempts to start the machine. but this ratio
was measured after one of the keypad read cables
was removed to be used for the status led. this
means the ratio could have been lower before,
when the machine was last in use for actual games